Title: 0에 가장 가까운 합 3, Time: 54ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 241115/0에 가장 가까운 합 3/sum-closest-to-zero-3.cpp
@@ -22,7 +22,7 @@ int main()
     {
         while (en - 1 > st)
         {
-            int dist = abs(board[en] + board[st]);
+            long long dist = abs(board[en] + board[st]);
             if (dist > abs(board[en - 1] + board[st]))
                 en--;
             else
